BL-5970 Consolidate Motion tool
* conflate Motion and MotionTool directories
