+++ src/BloomBrowserUI/bookEdit/toolbox/motion/motionTool.tsx
